requirements.txt
# You'll need Python3 to work outside the container.
# You'll also need suitable installations of Packer and Terraform.

ansible == 2.7.5
awscli == 1.16.76
boto == 2.49.0
boto3 == 1.9.66
jinja2 == 2.10
matildapeak-yacker == 2018.5
passlib == 1.7.1
pyOpenSSL == 18.0.0
pyyaml == 3.13
flake8 == 3.6.0
sphinx == 1.8.2
sphinx_rtd_theme == 0.4.2
